Add support for including or excluding files from tarballs and ZIPs
In order to support our new, submodule based distribution strategy for
Halyard, we need to be able to include or exclude our libs direcotry
when building tarballs.  This patch adds the ability to specify a list
of files in a tarball, and/or a list of excludes.  It also adds the
ability to specify a list of excludes when building ZIP files.

Also, we exclude any version control files.  tar includes a
--exclude-vcs that will automatically exclude .git, .gitignore, CVS,
.cvsignore, .svn, and various other version control files.  For the
make_zipfile command, we manually exclude .git and .gitignore files.