Update server installer: Refactor tests to eliminate common code
We had two tests of the update server installer; one for the base case
of doing an update.  They shared a lot of code.  This patch splits out
the shared code, and also splits out the tests for the base case into
a few smaller tests that are more specific.