Update server installer: Keep log of all changes to spec files
We would like to keep a log of all changes to our staging.spec and
release.spec files, so that we can more easily audit them, restore
previous versions if we need to, etc.  This patch adds a
staging.spec.log and release.spec.log, which contain lines from oldest
to newest in the following format:

old-build-id new-build-id username datetime notes

Thus, each line contains both the build ID of the file being replaced,
and the build ID of the new file (if the file is new, '<null>' is used
in place of the old-build-id).  The username is intended to be the
best guess at the username of the person pushing this version;
currently it is the username of the current user, though that will
likely not be useful if this is a shared account.  It includes a date
and time, in RFC 3339 (ISO 8601) format, and any notes associated with
this change (can be a version number if this was done while tagging a
release, or can be an explanation for why we're rolling back, or may
be empty).