project_config.rb: Add support for [sections], and better errors
For the update release scripts, I had added a project.conf for storing
configuration values needed to ssh in to the server and set up the
environment properly.  These had been a simple key=value format,
intended to be a subset of the relatively common INI file format.

This patch adds support for a larger subset of the INI file format,
including [sectons] and blank lines.  It also adds explicit errors
on common INI syntax that it doesn't understand, such as quoted
strings, comments, and \ escapes.

After reading a section heading, the parser will treat all keys
until the next section heading as if they were prefixed by the
section name followed by a ".".