project_config.rb: Add new parser for project configuration
In order to support project specific configuration, for such things as
ssh servers and usernames to be used for releasing updates, this patch
introduces a project config file, parsed by project_config.rb.  This
is a fairly simple file consisting of lines of "key=value" pairs, and
must be at config/project.conf.  This could probably be generalized
further, but at this stage, the simple version is sufficient for our
needs.