Add tools for staging and releasing updates on server
Our build script pushes updates to the update server, but the last
step of the process, moving an update from staging to release, has so
far been manual.  This patch automates this process, based on several
of the preceding patches.

This patch adds a new script, tools/release_update.rb, which is
intended to be run from within a project.  It acts somewhat like the
build script update installer build process, in that it copies the
entire build script library up to the server and runs scripts from it.
Depending on its arguments, it runs either
tools/release_update_from_staging.rb, or tools/stage_update.rb, both
of which are thin wrappers around equivalent functionality in
update_server_installer.rb.

This gets its configuration from the new config/project.conf that is
read by project_config.rb.  Some of the project-specific configuration
variables in the build script should probably be factored out into
calling the ProjectConfig parser.