Add support for Inno Setup Excludes containing \
In order to exclude some nested directories from IML libs in our programs,
we need to be able to match against more than one path component.  This
patch adds support for exclusions with multiple components, as documented
in the Inno Setup documentation[1], with the exact details worked out
by experimentation.

[1]: http://www.jrsoftware.org/ishelp/topic_filessection.htm