inno_setup.rb: Allow the same file to be installed to multiple locations
After the big Halyard reorganization, when we update programs from
before the reorg, we may be left with various broken shortcuts from
old installs.  To fix this, we would like to simply copy the same
source file to two different destinations in the installer and
updater.

This is supported by Inno Setup, but inno_setup.rb represented files
to install as a hash table with sources mapping to destinations, which
prevented the same source file from mapping to multiple destination
files.  This patch changes it to instead an array of FileCopy structs,
with a source and dest.